fix(artifacts): moved transient error codes to a new file
