+++ workflow/artifacts/s3/error_codes.go
@@ -1,0 +1,14 @@
+package s3
+
+// s3TransientErrorCodes is a list of S3 error codes that are transient (retryable)
+// Reference: https://github.com/minio/minio-go/blob/92fe50d14294782d96402deb861d442992038109/retry.go#L90-L102
+var s3TransientErrorCodes = []string{
+	"InternalError",
+	"RequestTimeout",
+	"Throttling",
+	"ThrottlingException",
+	"RequestLimitExceeded",
+	"RequestThrottled",
+	"InternalError",
+	"SlowDown",
+}
