+++ workflow/artifacts/s3/s3.go
@@ -31,19 +31,6 @@ type ArtifactDriver struct {
 	Context     context.Context
 }
 
-// s3TransientErrorCodes is a list of S3 error codes that are transient (retryable)
-// Reference: https://github.com/minio/minio-go/blob/92fe50d14294782d96402deb861d442992038109/retry.go#L90-L102
-var s3TransientErrorCodes = []string{
-	"InternalError",
-	"RequestTimeout",
-	"Throttling",
-	"ThrottlingException",
-	"RequestLimitExceeded",
-	"RequestThrottled",
-	"InternalError",
-	"SlowDown",
-}
-
 var (
 	_            artifactscommon.ArtifactDriver = &ArtifactDriver{}
 	defaultRetry                                = wait.Backoff{Duration: time.Second * 2, Factor: 2.0, Steps: 5, Jitter: 0.1}
